requirements.txt
openai
tk
Pillow==9.5.0
